+++ core/model/auth.go
@@ -334,6 +334,7 @@ type ServiceAccount struct {
 	Organization *Organization
 
 	Permissions []Permission
+	Scopes      []authorization.Scope
 	FirstParty  bool
 
 	Credentials []ServiceAccountCredential
@@ -349,6 +350,15 @@ func (s ServiceAccount) GetPermissionNames() []string {
 		permissions[i] = permission.Name
 	}
 	return permissions
+}
+
+// GetScopeStrings returns all names of scopes granted to this account
+func (s ServiceAccount) GetScopeStrings() []string {
+	scopes := make([]string, len(s.Scopes))
+	for i, scope := range s.Scopes {
+		scopes[i] = scope.String()
+	}
+	return scopes
 }
 
 // AppOrgPair represents an appID, orgID pair entity